requirements.txt
ibis
python-dotenv
inquirer
pandas
sqlalchemy
psycopg2
tqdm
